python-backend/requirements-windows.txt
# FastAPI Backend for Voice Assistant Health System - Windows Compatible
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
websockets>=12.0

# Voice Assistant Dependencies
google-generativeai>=0.3.0
speechrecognition>=3.10.0
python-dotenv>=1.0.0
gtts>=2.3.0
playsound==1.2.2

# Real-time Audio Processing (Windows compatible alternatives)
# webrtcvad>=2.0.10  # Commented out - requires Visual C++
numpy>=1.24.0
# wave>=0.0.2  # Commented out - not needed (built-in module)

# Additional FastAPI dependencies
aiofiles>=23.0.0

# Utility libraries
requests>=2.28.0

# Development dependencies (optional)
pytest>=7.0.0
httpx>=0.24.0

# Windows-specific audio alternatives
# pyaudio  # Commented out - can be problematic on Windows
# Use portaudio or sounddevice as alternatives if needed